+++ src/vscp/common/vscp-bootdevice-vscp.h
@@ -175,7 +175,7 @@ public:
         MDF is not the same as the one read from the remote device.
       @return VSCP_ERROR_SUCCESS on success.
   */
-  int deviceInit(cguid &ourguid, uint8_t devicecode, bool bAbortOnFirmwareCodeFail = false);
+  int deviceInit(cguid &ourguid, uint16_t devicecode, bool bAbortOnFirmwareCodeFail = false);
 
   /*!
       Perform the actual firmware load process
